pyproject.toml
[project]
name = "maang-stocks"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Paul Hissen", email = "paul.hissen@gmail.com"},
]
dependencies = [
    "fastapi>=0.115.6",
    "pymongo>=4.10.1",
    "jwt>=1.3.1",
    "uvicorn>=0.33.0",
    "python-jose>=3.3.0",
    "invoke>=2.2.0",
    "commitizen>=4.1.0",
    "python-dotenv>=1.0.1",
    "axios>=0.4.0",
    "motor>=3.5.1",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$version"
bump_message = "release: $new_version"
version_scheme = "pep440"
version_provider = "pep621"
update_changelog_on_bump = true
annotated_tag = true
